moving check running
