+++ cpaeds/utils.py
@@ -73,24 +73,6 @@ def check_finished(settings_loaded):
 
     return run_complete, len(omd_list)
 
-def check_job_running(user,status,run,dir):
-    with open('running_jobs.out', 'w+') as outfile:
-        exe = subprocess.run(
-                ['squeue', '-u', user],
-                check=True,
-                stdout=outfile,
-                capture_output= False,
-                text=True,
-            )
-    exe.check_returncode()
-    with open('running_jobs.out', 'r') as paths:
-        for line in paths:
-            if f"{dir} " in line:
-                line = line.split()
-                status[f"run_{run+1}"] = line
-                return True
-    os.remove('running_jobs.out')   
-
 def copy_lib_file(destination,name):
     path = os.path.abspath(os.path.join(os.path.dirname(__file__), f"data/{name}"))
     if os.path.exists(f"{destination}/{name}"):
